apps/user-service/requirements.txt
fastapi
uvicorn[standard]
pyjwt
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
